Safari-611.1.10/LayoutTests/fast/images/image-usemap-parsing-expected.txt


Tests that usemap attribute parsing logic.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

usemap=#navmap
PASS areaWasClicked is true

usemap=navmap
PASS areaWasClicked is false

usemap=#navMap
PASS areaWasClicked is false

usemap=foo#navmap
PASS areaWasClicked is true

usemap=foo#navmap2
PASS areaWasClicked is false

usemap=navmap#
PASS areaWasClicked is false

usemap=navmap#bar
PASS areaWasClicked is false

usemap=#
PASS areaWasClicked is false

usemap=
PASS areaWasClicked is false

PASS successfullyParsed is true

TEST COMPLETE
